Ground work for `wgpu-core-remote-types`
