+++ Cargo.toml
@@ -22,6 +22,7 @@ members = [
     "wgpu-core",
     "wgpu-core/platform-deps/*",
     "wgpu-core-remote",
+    "wgpu-core-remote-types",
     "wgpu-hal",
     "wgpu-info",
     "wgpu-naga-bridge",
@@ -93,6 +94,8 @@ wgpu = { version = "30.0.0", path = "./wgpu", default-features = false, features
     "static-dxc",
 ] }
 wgpu-core = { version = "30.0.0", path = "./wgpu-core" }
+wgpu-core-remote = { version = "30.0.0", path = "./wgpu-core-remote" }
+wgpu-core-remote-types = { version = "30.0.0", path = "./wgpu-core-remote-types" }
 wgpu-hal = { version = "30.0.0", path = "./wgpu-hal" }
 wgpu-naga-bridge = { version = "30.0.0", path = "./wgpu-naga-bridge" }
 wgpu-sync = { version = "30.0.0", path = "./wgpu-sync", default-features = false }
