+++ wgpu-core-remote-types/src/lib.rs
@@ -1,0 +1,7 @@
+extern crate alloc;
+
+pub type Index = u32;
+pub type Epoch = u32;
+
+pub mod id;
+pub mod identity;
